move session sharing config settings to its own react component
